\name{autoconf-package}
\alias{autoconf-package}
\alias{autoconf}
\docType{package}
\title{\packageTitle{autoconf}}
\description{\packageDescription{autoconf}}
\section{Package Content}{\packageIndices{autoconf}}
\author{\packageAuthor{autoconf}}
\section{Maintainer}{\packageMaintainer{autoconf}}
\keyword{package}
